requirements.txt
pygame
mysql-connector-python
matplotlib>=3.7.0
